# Improve performance of `--d3-flamegraph` option