--- a/lib/stackprof/report.rb
+++ b/lib/stackprof/report.rb
@@ -95,51 +95,10 @@ module StackProf
       print_flamegraph(f, skip_common, true)
     end
 
-    StackCursor = Struct.new(:raw, :idx, :length) do
-      def weight
-        @weight ||= raw[1 + idx + length]
-      end
-
-      def [](i)
-        if i >= length
-          nil
-        else
-          raw[1 + idx + i]
-        end
-      end
-
-      def <=>(other)
-        i = 0
-        while i < length && i < other.length
-          if self[i] != other[i]
-            return self[i] <=> other[i]
-          end
-          i += 1
-        end
-
-        return length <=> other.length
-      end
-    end
-
     def print_flamegraph(f, skip_common, alphabetical=false)
       raise "profile does not include raw samples (add `raw: true` to collecting StackProf.run)" unless raw = data[:raw]
 
-      stacks = []
-      max_x = 0
-      max_y = 0
-
-      idx = 0
-      loop do
-        len = raw[idx]
-        break unless len
-        max_y = len if len > max_y
-
-        stack = StackCursor.new(raw, idx, len)
-        stacks << stack
-        max_x += stack.weight
-
-        idx += len + 2
-      end
+      stacks, max_x, max_y = flamegraph_stacks(raw)
 
       stacks.sort! if alphabetical
 
@@ -150,7 +109,7 @@ module StackProf
         x = 0
 
         stacks.each do |stack|
-          weight = stack.weight
+          weight = stack.last
           cell = stack[y] unless y == stack.length-1
 
           if cell.nil?
@@ -189,6 +148,24 @@ module StackProf
         end
       end
       f.puts '])'
+    end
+
+    def flamegraph_stacks(raw)
+      stacks = []
+      max_x = 0
+      max_y = 0
+      idx = 0
+
+      while len = raw[idx]
+        idx += 1
+        max_y = len if len > max_y
+        stack = raw.slice(idx, len+1)
+        idx += len+1
+        stacks << stack
+        max_x += stack.last
+      end
+
+      return stacks, max_x, max_y
     end
 
     def flamegraph_row(f, x, y, weight, addr)
@@ -231,19 +208,7 @@ module StackProf
     def print_d3_flamegraph(f=STDOUT, skip_common=true)
       raise "profile does not include raw samples (add `raw: true` to collecting StackProf.run)" unless raw = data[:raw]
 
-      stacks = []
-      max_x = 0
-      max_y = 0
-      idx = 0
-
-      while len = raw[idx]
-        idx += 1
-        max_y = len if len > max_y
-        stack = raw.slice(idx, len+1)
-        idx += len+1
-        stacks << stack
-        max_x += stack.last
-      end
+      stacks, * = flamegraph_stacks(raw)
 
       # d3-flame-grpah supports only alphabetical flamegraph
       stacks.sort!