@mame 2021-05-07
Looks good. I copied the code from `print_flamegraph`, so the improvement should be applied to it as well. Or the code should be factored out to a method.