@pocke 2021-05-07
@mame Thanks for your comment! I didn't know that `print_flamegraph` method has the same logic.



`print_flamegraph` has been optimized since https://github.com/tmm1/stackprof/pull/136, but the optimization has a bit of a different approach. 
I extracted my approach as a method because it is faster than the other.

So this change makes `--flamegraph` option faster also.
Previously `stackprof --flamegraph stackprof-out` takes 0.88 sec with the Fibonacci example but now it takes 0.60 sec.



The difference for `--d3-flamegraph` is larger than `--flamegraph`'s one.
If `--d3-flamegraph` uses the `StackCursor` approach, it takes 5.46 sec. However it takes 1.03 sec on this pull request.